Fix: audio cache would overflow over int sized count and position differences

+++ avs_core/core/cache.h
@@ -52,7 +52,7 @@ private:
   Device* device;
   std::mutex& CacheGuardMutex; // just reference!
 
-  void FillAudioZeros(void* buf, int start_offset, int count);
+  void FillAudioZeros(void* buf, size_t start_offset, size_t count);
 
 public:
 #ifdef _DEBUG
